Mlops/requirements.txt
pandas
numpy
openpyxl
python-dotenv
db-dtypes
jupyter
notebook
ipykernel
seaborn
ipywidgets
scikit-learn
xgboost
imbalanced-learn
pydantic
fastapi
uvicorn 
streamlit
evidently
google-cloud-bigquery
db-dtypes